tasks/lms/requirements.txt
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyMySQL==1.0.2
PyYAML==6.0
Werkzeug==2.0.2
gunicorn==19.9.0

